Split unit tests file
